Add DOOM Guy
First #book of the year #read: "Doom Guy: Life in First Person" by #JohnRomero: I enjoyed this a lot, especially because it's quite personally and somewhat naive (in a positive way) written. Great insights, and as he writes at the end: Focussing on the positive aspects.

Also, there seem to be humble people in tech still. I guess it depends a big chunk on your upbringing, if you ever actually worked for your income.

Good stuff.