Gelesen / Read: Why Has Nobody Told Me This before.
This came to us as a gift and me first reaction was a bit of a smirk, but eventually I grabbed toe copy to and read it fully.
For large extends it feels to me like a book on the "Litany Against Fear" and a set of tools to avoid the Gom Jabbar to sting.
This sounds maybe more harsh than intended, I did actually like reading it.
In an unexpected turn of events, I do feel that I am applying a lot of the recommendations in the last 5 or so years in my own life, which is kinda reassuring.

The thing I really liked most in the book is the concept of the "Feeling wheel": A word wheel aimed to avoid using only few words to describe everything as "happy", "aweful", "freightening". I will certainly think more about that in the future to come. It's true, the words that we have shape what we are able to think.

The mindfulness thing on it's own it's not my cup of tea, or maybe it is, but not in the way describe here.