Gelesen / Read: "Es ist okay, wütend auf den Kapitalismus zu sein"
This book by Sanders, published in 2023, is very much written with a focus on US politics, but for a European / German person interesting nevertheless.
We often enough observe that Germany is lagging behind US only a few years.
The book was written before the 2024 election, obviously, and I wonder how Sanders feels about it today?
Everything and more he has been warning about is coming into reality in the last weeks.
It's a loud call to pick up a site: Are you on the side of the working class or on the site of the oligarchs shredding society and the planet into pieces for their own endulgement.
Sanders makes a clear connection why picking the side of the working class is ultimatively better for society as a whole, and the planet.
As German, it's frightening to me how bad the US health system actually is, how little effort goes into public school system etc.
The last chapter directly adresses the two party system and why the democrats will fail if they don't pick a side and we observe this in Germany already:
If you keep on insisting that you are "just better" than the obviously dangerous other parties, who is already stirring anxiety, fear and hate by blaming scape goats instead of adressing underlying problems, you will lose to the other party: Voters will go for "the original", regardless if it's a worse choice.

Recommend for getting a great view into that society.