Gelesen / Read: Sarah Wynn-Williams, Careless People: A story of where I used to work
There's a lot of hype around that book, and if you never wasted a thought how things might are in a company that is build on exploitation of their users to begin with, you might be in for a surprise.
The most interesting bits and pieces is actually the testament of how Mark Zuckerberg changed from a nerd persona to the current Zuck or nothing one, from "i don't do politicians" (just like some engineers never really want to hear about a domain) to "i am better than your next average president" (again, same hybris like some engineers… just because you know some programming, you are not bound to solve all the problems better).

If only half of it is acurate, the book also sheds a light on how terrible working conditions are in US tech companies.
Sure, you have a good chance to become a multi-millionar before you are 30, but to what price?

Especially in that light I have to say - and I do pity the author - that she is very complicit to what she have bore witness to at Facebook.
7 years plus and year after year thinking only about an exit? C'm on, that's hard to believe it would have been that necessary.
Or, trying to be less judgemental: Everyone might be vulnerable to a golden handcuff.