Gelesen / Read: Dan Simmons - Die Hyperion Gesänge
What a ride. My dear colleague Tom G. recommended this to me and after reading, I am surprised
i didn't hear about it before. I like that Space-Opera theme spawned over aeons.
Having said that: Different than anything else I read in that genre. Frank Herberts Dune is wild
in the later books, but this is odd of this world wild. Probably give it a reread in a year or two
to capture all the details.

Wrt artifical intelligence its funny/sad how accurate the book foresees change in behaviour in humans,
taking the comfort of things for the exchange of free will.